# contracts/oar-openapi.yaml
openapi: 3.1.0
info:
  title: Organization Autorunner HTTP API
  version: 0.3.0
  description: |
    Canonical workspace HTTP contract for Organization Autorunner.

    Topics and cards are the primary durable resources. Threads remain backing
    infrastructure surfaced through read-only inspection and workspace views.
servers:
  - url: http://127.0.0.1:8000
...
paths:
...
  /threads:
    get:
      operationId: listThreads
      summary: List backing threads
      tags: [threads]
      x-oar-surface: canonical
      x-oar-command-id: threads.list
      x-oar-cli-path: threads list
      x-oar-why: Inspect backing infrastructure threads without making them the primary planning noun.
      x-oar-input-mode: none
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ threads }`."
      x-oar-error-codes: [auth_required, invalid_token]
      x-oar-concepts: [threads, inspection]
      x-oar-stability: beta
      responses:
        '200':
          description: Thread list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadsResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /threads/{thread_id}:
    get:
      operationId: getThread
      summary: Inspect backing thread
      tags: [threads]
      x-oar-surface: canonical
      x-oar-command-id: threads.inspect
      x-oar-cli-path: threads inspect
      x-oar-why: Resolve one backing thread for low-level inspection and diagnostics.
      x-oar-input-mode: none
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ thread }`."
      x-oar-error-codes: [auth_required, invalid_token, not_found]
      x-oar-concepts: [threads, inspection]
      x-oar-stability: beta
      parameters:
        - $ref: '#/components/parameters/ThreadID'
      responses:
        '200':
          description: Thread payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
  /docs:
    get:
      operationId: listDocuments
      summary: List documents
      tags: [docs]
      x-oar-surface: canonical
      x-oar-command-id: docs.list
      x-oar-cli-path: docs list
      x-oar-why: Scan canonical document lineages.
      x-oar-input-mode: none
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ documents }`."
      x-oar-error-codes: [auth_required, invalid_token]
      x-oar-concepts: [docs]
      x-oar-stability: beta
      responses:
        '200':
          description: Document list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentsResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
    post:
      operationId: createDocument
      summary: Create document
      tags: [docs]
      x-oar-surface: canonical
      x-oar-command-id: docs.create
      x-oar-cli-path: docs create
      x-oar-why: Create a canonical document lineage anchored to a typed subject ref.
      x-oar-input-mode: json-body
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ document, revision }`."
      x-oar-error-codes: [auth_required, invalid_request, invalid_token]
      x-oar-concepts: [docs, write]
      x-oar-stability: beta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentRequest'
      responses:
        '201':
          description: Document created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentWithRevisionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /docs/{document_id}:
    get:
      operationId: getDocument
      summary: Get document
      tags: [docs]
      x-oar-surface: canonical
      x-oar-command-id: docs.get
      x-oar-cli-path: docs get
      x-oar-why: Resolve a document lineage and its current head revision.
      x-oar-input-mode: none
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ document, revision }`."
      x-oar-error-codes: [auth_required, invalid_token, not_found]
      x-oar-concepts: [docs]
      x-oar-stability: beta
      parameters:
        - $ref: '#/components/parameters/DocumentID'
      responses:
        '200':
          description: Document payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentWithRevisionResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
  /artifacts/{artifact_id}:
    get:
      operationId: getArtifact
      summary: Get artifact metadata
      tags: [artifacts]
      x-oar-surface: canonical
      x-oar-command-id: artifacts.get
      x-oar-cli-path: artifacts get
      x-oar-why: Resolve immutable artifact metadata referenced from timelines and packets.
      x-oar-input-mode: none
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ artifact }`."
      x-oar-error-codes: [auth_required, invalid_token, not_found]
      x-oar-concepts: [artifacts]
      x-oar-stability: beta
      parameters:
        - $ref: '#/components/parameters/ArtifactID'
      responses:
        '200':
          description: Artifact payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  responses:
    ErrorResponse:
      description: Error payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
...
    ThreadID:
      in: path
      name: thread_id
      required: true
      schema: { type: string }
    DocumentID:
      in: path
      name: document_id
      required: true
      schema: { type: string }
...
    ArtifactID:
      in: path
      name: artifact_id
      required: true
      schema: { type: string }
...
  schemas:
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
      required: [code, message]
...
    TypedRef:
      type: string
      description: Typed ref string such as `topic:roadmap` or `artifact:art_123`.
    Provenance:
      type: object
      properties:
        sources:
          type: array
          items: { type: string }
        notes: { type: string }
        by_field:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
      required: [sources]
...
    Artifact:
      type: object
      properties:
        id: { type: string }
        kind: { type: string }
        content_type: { type: string }
        content_hash: { type: string }
        refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        summary: { type: string }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
      required: [id, kind, content_type, content_hash, refs, created_at, created_by]
    Thread:
      type: object
      properties:
        id: { type: string }
        topic_ref: { $ref: '#/components/schemas/TypedRef' }
        title: { type: string }
        status:
          type: string
          enum: [active, archived]
        priority:
          type: string
          enum: [p0, p1, p2, p3]
        tags:
          type: array
          items: { type: string }
        latest_event_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
        updated_at: { type: string, format: date-time }
        updated_by: { type: string }
        provenance: { $ref: '#/components/schemas/Provenance' }
      required: [id, title, status, priority, tags, created_at, created_by, updated_at, updated_by, provenance]
...
    Document:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        subject_ref: { $ref: '#/components/schemas/TypedRef' }
        head_revision_ref: { $ref: '#/components/schemas/TypedRef' }
        state: { type: string }
        refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
        updated_at: { type: string, format: date-time }
        updated_by: { type: string }
        provenance: { $ref: '#/components/schemas/Provenance' }
      required:
        [id, title, subject_ref, head_revision_ref, state, refs, created_at, created_by, updated_at, updated_by, provenance]
    DocumentRevision:
      type: object
      properties:
        revision_id: { type: string }
        document_ref: { $ref: '#/components/schemas/TypedRef' }
        prev_revision_ref: { $ref: '#/components/schemas/TypedRef' }
        artifact_ref: { $ref: '#/components/schemas/TypedRef' }
        revision_number: { type: integer }
        summary: { type: string }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
      required: [revision_id, document_ref, artifact_ref, revision_number, created_at, created_by]
...
    DocumentCreateInput:
      type: object
      properties:
        title: { type: string }
        subject_ref: { $ref: '#/components/schemas/TypedRef' }
        body_markdown: { type: string }
        summary: { type: string }
        refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        provenance: { $ref: '#/components/schemas/Provenance' }
      required: [title, subject_ref, body_markdown, refs, provenance]
...
    CreateDocumentRequest:
      type: object
      properties:
        document: { $ref: '#/components/schemas/DocumentCreateInput' }
      required: [document]
...
    ThreadResponse:
      type: object
      properties:
        thread: { $ref: '#/components/schemas/Thread' }
      required: [thread]
    ThreadsResponse:
      type: object
      properties:
        threads:
          type: array
          items: { $ref: '#/components/schemas/Thread' }
      required: [threads]
...
    DocumentsResponse:
      type: object
      properties:
        documents:
          type: array
          items: { $ref: '#/components/schemas/Document' }
      required: [documents]
    DocumentWithRevisionResponse:
      type: object
      properties:
        document: { $ref: '#/components/schemas/Document' }
        revision: { $ref: '#/components/schemas/DocumentRevision' }
      required: [document, revision]
...
    ArtifactResponse:
      type: object
      properties:
        artifact: { $ref: '#/components/schemas/Artifact' }
      required: [artifact]
...